fix build warnings
