+++ docs/fundamentals/runtime-libraries/snippets/System.Data/DataTable/Overview/cs/project.csproj
@@ -2,7 +2,8 @@
 
   <PropertyGroup>
     <OutputType>Library</OutputType>
-    <TargetFrameworks>net8</TargetFrameworks>
+    <TargetFramework>net4.8</TargetFramework>
+    <UseWindowsForms>true</UseWindowsForms>
   </PropertyGroup>
 
- </Project>
+</Project>
