+++ docs/fundamentals/runtime-libraries/system-iasyncdisposable.md
@@ -1,6 +1,6 @@
 ---
 title: System.IAsyncDisposable interface
-description: Learn about the System.IDisposable interface.
+description: Learn about the System.IAsyncDisposable interface.
 ms.date: 12/31/2023
 ---
 # System.IAsyncDisposable interface
